org-auto-repeat-maybe: Set up org-add-log-note hook after any prompts
Prior to this change, when `org-log-repeat` is `'time`, invoking
`org-todo` caused the following message when the "%d repeater
intervals..." prompt appears:

TK

Additionally, the state transition is not logged properly when this
happens.

This commit changes the `post-command-hook` to be set up later, so
that the `y-or-n-p` does not cause it to be triggered too early.